fix(core): circular imports of SDK_VERSION
