+++ packages/core/accessibility/index.android.ts
@@ -1,6 +1,6 @@
 import * as Application from '../application';
 import { Trace } from '../trace';
-import { SDK_VERSION } from '../utils';
+import { SDK_VERSION } from '../utils/constants';
 import type { View } from '../ui/core/view';
 import { GestureTypes } from '../ui/gestures';
 import { notifyAccessibilityFocusState } from './accessibility-common';
