+++ packages/core/utils/constants.android.ts
@@ -1,0 +1,1 @@
+export const SDK_VERSION = android.os.Build.VERSION.SDK_INT;
