+++ packages/core/utils/constants.d.ts
@@ -1,0 +1,1 @@
+export const SDK_VERSION: number;
