+++ packages/core/utils/constants.ios.ts
@@ -1,0 +1,1 @@
+export const SDK_VERSION = parseFloat(UIDevice.currentDevice.systemVersion);
